renamed libssh2_sftp_packet_requirev to plain sftp_packet_requirev since it is private in this file only and a shorter name is nicer
Removed a "flush" of the data in sftp_packet_requirev() that now seems to have
made SFTP operations a lot more reliable. It didn't make much sense to have it
there but if someone can present a reason for one I figure we should carefully
investigate one and only do it conditionally where/when needed.