libssh2_channel_read_ex: 0 means EOF
Until now, libssh2_channel_read_ex was returning 0 for two cases:

1) At EOF

2) After the data transport layer had produced data, but not for the
current channel.

And it was not possible to discern between the two conditions in any
way (note that libssh2_channel_eof would not return true when data is
still queued for the the other channel stream - !ext).

This patch, changes that behavior such that now, only on the first
case (at EOF) a zero is returned.

And the second case is converted to an EAGAIN.