scp_recv: detect premature EOF properly
Before this patch, "libssh2_channel_eof" was used to check for a
premature EOF but that condition could be masked by data queued in the
incoming extended stream. That resulted in "scp_recv" returning the
scp channel object in an incomplete state.

The change introduced in #145 which ensures that libssh2_read_ex
returns 0 only on EOF, makes this logic unnecesary.