scp_recv: detect premature EOF properly
Before this patch, "libssh2_channel_eof" was used to check for a
premature EOF but that condition could be masked by data queued in the
incoming extended stream. That resulted in "scp_recv" returning the
scp channel object in an incomplete state.

The change introduced in #104 made this logic unnecesary. Now, when
read returns 0, it means EOF always.