_libssh2_channel_read: refactoring
The logic used to check when a packet has data for the requested
channel/stream was quite convoluted. Now it is much simpler.

Also, error handling has been delayed until after looking for data in
the packet list. Even if a fatal error pops up, data may be queued
there.